+++ doc/whats_new.rst
@@ -23,6 +23,9 @@ Changes
  - cmap is now edge_cmap
  - threshold is now edge_threshold
 
+- plotting.view_markers and plotting.view_connectome can accept different marker
+  sizes for each node / marker.
+
 - plotting.view_markers() default marker color is now 'red', consistent with add_markers().
 - plotting.view_markers() parameter names are consistent with add_markers():